// Enums/Messages.ts
enum Messages {
  USERNAME_REQUIRED = 'Username is required',
  USERNAME_MUSTBE_STRING = 'Username must be a string',
  USERNAME_VALID_LENGTH = 'Username must be longer than 2 characters',
  CLASSE_REQUIRED = 'Classe is required',
  CLASSE_MUSTBE_STRING = 'Classe must be a string',
  CLASSE_VALID_LENGTH = 'Classe must be longer than 2 characters',
  LEVEL_REQUIRED = 'Level is required',
  LEVEL_MUSTBE_NUMBER = 'Level must be a number',
  LEVEL_VALID = 'Level must be greater than 0',
  PASSWORD_REQUIRED = 'Password is required',
  PASSWORD_MUSTBE_STRING = 'Password must be a string',
  PASSWORD_VALID_LENGTH = 'Password must be longer than 7 characters',
  USER_PASSWORD = 'Username or password invalid',
  TKN_NTF = 'Token not found',
  TKN_INVALID = 'Invalid token',
  NAME_REQUIRED = 'Name is required',
  NAME_MUSTBE_STRING = 'Name must be a string',
  NAME_VALID_LENGTH = 'Name must be longer than 2 characters',
  AMOUNT_REQUIRED = 'Amount is required',
  AMOUNT_MUSTBE_STRING = 'Amount must be a string',
  AMOUNT_VALID_LENGTH = 'Amount must be longer than 2 characters',

}

export default Messages;